+++ language/en/info_acp_pbwowext.php
@@ -32,9 +32,8 @@ if (empty($lang) || !is_array($lang))
 
 $lang = array_merge($lang, array(
 	// Extension modules
-	'ACP_PBWOW3_CATEGORY'		=> 'PBWoW 3',
-	'ACP_PBWOW3_OVERVIEW'		=> 'Overview',
-	'ACP_PBWOW3_CONFIG'			=> 'Configuration',
+	'ACP_PBWOWEXT_CATEGORY'		=> 'PBWoW 3',
+	'ACP_PBWOWEXT_CONFIG'			=> 'Configuration',
 
 	// Common
 	'PBWOW_ACTIVE'				=> 'active',
@@ -47,9 +46,7 @@ $lang = array_merge($lang, array(
 
 	'LOG_PBWOW_CONFIG'			=> '<strong>Altered PBWoW settings</strong><br />&raquo; %s',
 
-
 	// OVERVIEW //
-
 	'PBWOW_OVERVIEW_TITLE'				=> 'PBWoW Extension Overview',
 	'PBWOW_OVERVIEW_TITLE_EXPLAIN'		=> 'Thank you for choosing PBWoW, hope you like it.',
 	'ACP_PBWOW_INDEX_SETTINGS'			=> 'General information',
@@ -63,42 +60,18 @@ $lang = array_merge($lang, array(
 	'PBWOW_LATEST_VERSION'				=> 'Latest version',
 	'PBWOW_EXT_VERSION'					=> 'Extension version',
 	'PBWOW_STYLE_VERSION'				=> 'Style version',
-	'PBWOW_VERSION_ERROR'				=> 'Unable to determine version!',
+	'PBWOW_LATEST_STYLE_VERSION'		=> 'Latest Style version',
+	'PBWOW_VERSION_ERROR'				=> 'Unable to check latest version!',
 	'PBWOW_CHECK_UPDATE'				=> 'Check <a href="http://www.avathar.be/bbdkp/index.php">avathar.be</a> to see if there are updates available.',
 
-	'PBWOW_CPF_CHECK'					=> 'Custom Profile Fields Check',
-	//'PBWOW_CPF_CREATE_LOCATION'			=> 'Create or enable this field via ACP > Users and Groups > Custom profile fields',
-	'PBWOW_CPF_LOAD_LOCATION'			=> 'Enable this via ACP > General > Board Configuration > Board Features',
-	'PBWOW_GAME_EXPLAIN'				=> 'The custom profile fields for this game are currently disabled.',
-
-	'PBWOW_BNETCHARS_CHECK'				=> 'Battle.net API character information functionality',
-	'PBWOW_CHARSDB_GOOD'				=> 'PBWoW characters table found (%s)',
-	'PBWOW_CHARSDB_BAD'					=> 'No PBWoW characters table found. Make sure that the table (%s) exists in your phpBB database.',
-	'PBWOW_CHARSDB_BAD_EXPLAIN'			=> 'The required PBWoW 3 Battle.net Characters database table should have been installed automatically when you installed the PBWoW extension. Please uninstall it, delete the data, and try installing it again.',
-	'PBWOW_CHARSDB_FLUSH'				=> 'Flush/clear the characters table',
-	'PBWOW_CHARSDB_FLUSH_EXPLAIN'		=> 'This will clear all the Battle.net character information stored in the DB. It will be retrieved again automatically when needed.',
-	'PBWOW_CURL_BAD'					=> 'Your server does not allow &quot;cURL&quot;!',
-	'PBWOW_CURL_BAD_EXPLAIN'			=> 'Check your server config, or contact your server host. Disable the Battle.net API until cURL is enabled!',
-
+	'PBWOW_DONATE_URL'                  => 'http://www.avathar.be/bbdkp/app.php/page/donate',
+	'PAYPAL_IMAGE_URL'                  => 'https://www.paypalobjects.com/webstatic/en_US/i/btn/png/silver-pill-paypal-26px.png',
+	'PAYPAL_ALT'                        => 'Donate using PayPal',
 	'PBWOW_DONATE'						=> 'Donate to PBWoW',
 	'PBWOW_DONATE_SHORT'				=> 'Make a donation to PBWoW',
 	'PBWOW_DONATE_EXPLAIN'				=> 'PBWoW is 100% free. It is a hobby project that I am spending my time and money on, just for the fun of it. If you enjoy using PBWoW, please consider making a donation. I would really appreciate it. No strings attached.',
 
-
-	// LEGACY CHECKS //
-
-	'PBWOW_LEGACY_CHECK'				=> 'PBWoW Legacy Check',
-
-	'PBWOW_LEGACY_CONSTANTS'			=> 'PBWoW Legacy Constants',
-	'PBWOW_LEGACY_CONSTANTS_EXPLAIN'	=> 'If detected, this means that there are still (partial) MODs of PBWoW v1 or v2 active! This could potentially lead to errors. That is why we strongly urge you to uninstall any active (PBWoW) MODs before upgrading to the latest phpBB version. Either that, or install a clean phpBB version and use the database update function of the phpBB installer.',
-	'PBWOW_LEGACY_DATABASE'				=> 'PBWoW Legacy Database(s)',
-	'PBWOW_LEGACY_DATABASE_EXPLAIN'		=> 'The config table of PBWoW v1 or v2 is still active. This is no problem, since PBWoW 3 does not interact with it. But you can drop/delete the table if you want (and are no longer using it).',
-
-	'PBWOW_LEGACY_NONE'					=> 'No obvious potentially problematic traces of older PBWoW versions were found. This is good.',
-
-
 	// CONFIG //
-
 	'PBWOW_CONFIG_TITLE'				=> 'PBWoW Configuration',
 	'PBWOW_CONFIG_TITLE_EXPLAIN'		=> 'Here you can choose some options to customize your PBWoW installation.',
 
@@ -133,24 +106,6 @@ $lang = array_merge($lang, array(
 	'PBWOW_VIDEOBG_ALLPAGES_EXPLAIN'	=> 'By default, PBWoW only loads the video backgrounds (if available) on <u>index.php</u> pages. You can enable them for all pages, but this may affect the browsing speed of your visitors (but in general not your server bandwidth, because they are cached locally). [only applies if video is enabled]',
 	'PBWOW_FIXEDBG'						=> 'Fixed background position',
 	'PBWOW_FIXEDBG_EXPLAIN'				=> 'Fixing the background position will prevent it from scrolling along with the rest of the content. Keep in mind that some lower resolution devices will have no option to see the entire background image.',
-
-	'PBWOW_AVATARS'						=> 'Gaming Avatars',
-	'PBWOW_AVATARS_ENABLE'				=> 'Enable board-wide gaming avatars (and icons)',
-	'PBWOW_AVATARS_ENABLE_EXPLAIN'		=> 'If enabled, your board will display a generated gaming avatar (based on profile field entries) if the user has no custom avatar configured.',
-	'PBWOW_AVATARS_PATH'				=> 'Gaming avatars path',
-	'PBWOW_AVATARS_PATH_EXPLAIN'		=> 'Path under your phpBB root directory where the gaming avatars are stored, e.g. <samp>images/avatars/gaming</samp>.<br />Character icons also require this path to be set.',
-	'PBWOW_SMALLRANKS_ENABLE'			=> 'Use small rank-images',
-	'PBWOW_SMALLRANKS_ENABLE_EXPLAIN'	=> 'Enable this if you wish to use small rank-images that overlay the avatar (as it does on PBWoW.com). Don&#39;t enable this if you are using larger rank-images.',
-
-	'PBWOW_BNET_APIKEY'					=> 'Battle.net API Key',
-	'PBWOW_BNET_APIKEY_EXPLAIN'			=> 'Enter your Battle.net game API key. If you don&#39;t have one, get one by creating a <a href="https://dev.battle.net/member/register">Mashery account</a>.',
-	'PBWOW_BNETCHARS'					=> 'Battle.net Character Information',
-	'PBWOW_BNETCHARS_ENABLE'			=> 'Enable Battle.net API character information',
-	'PBWOW_BNETCHARS_ENABLE_EXPLAIN'	=> 'Enable this feature to use the Battle.net API to retrieve character information (when available), for use in user profiles. The <u>Gaming Avatars</u> setting must be enabled to display Battle.net avatars!',
-	'PBWOW_BNETCHARS_CACHETIME'			=> 'Cache time-to-live',
-	'PBWOW_BNETCHARS_CACHETIME_EXPLAIN'	=> 'Sets the time-to-live (in seconds) of cached character information after it has been retrieved from the Battle.net API. You can change this to update character information more or less frequently. 86400 = 24h',
-	'PBWOW_BNETCHARS_TIMEOUT'			=> 'API query time-out',
-	'PBWOW_BNETCHARS_TIMEOUT_EXPLAIN'	=> 'Sets the time-out interval (in seconds) of Battle.net API requests. Basically meaning the maximum time that the script will wait for Battle.net to respond. Increase this if you think that (correct) data is not being received on time, but page load time can increase!',
 
 	'PBWOW_ADS_INDEX'					=> 'Index Advertisement Block',
 	'PBWOW_ADS_INDEX_ENABLE'			=> 'Enable index advertisement',