+++ language/fr/info_acp_pbwowext.php
@@ -32,9 +32,8 @@ if (empty($lang) || !is_array($lang))
 
 $lang = array_merge($lang, array(
 	// Extension modules
-	'ACP_PBWOW3_CATEGORY'		=> 'PBWoW 3',
-	'ACP_PBWOW3_OVERVIEW'		=> 'Vue d’ensemble',
-	'ACP_PBWOW3_CONFIG'			=> 'Configuration',
+	'ACP_PBWOWEXT_CATEGORY'		=> 'PBWoW 3',
+	'ACP_PBWOWEXT_CONFIG'		=> 'Configuration',
 
 	// Common
 	'PBWOW_ACTIVE'				=> 'actif',
@@ -49,7 +48,6 @@ $lang = array_merge($lang, array(
 
 
 	// OVERVIEW //
-
 	'PBWOW_OVERVIEW_TITLE'				=> 'Vue d’ensemble de l’extension PBWoW',
 	'PBWOW_OVERVIEW_TITLE_EXPLAIN'		=> 'Merci d’utiliser PBWoW, nous espérons que cela vous plaira.',
 	'ACP_PBWOW_INDEX_SETTINGS'			=> 'Informations générales',
@@ -63,39 +61,16 @@ $lang = array_merge($lang, array(
 	'PBWOW_LATEST_VERSION'				=> 'Dernière version',
 	'PBWOW_EXT_VERSION'					=> 'Version de l’extension',
 	'PBWOW_STYLE_VERSION'				=> 'Version du style',
+	'PBWOW_LATEST_STYLE_VERSION'		=> 'Dernière version du Style',
 	'PBWOW_VERSION_ERROR'				=> 'Impossible de déterminer la version !',
-	'PBWOW_CHECK_UPDATE'				=> 'Vérifier <a href="http://www.avathar.be/bbdkp/index.php">avathar.be</a> pour voir s’il y a des mises à jour disponibles.',
+	'PBWOW_CHECK_UPDATE'				=> 'Vérifie <a href="http://www.avathar.be/bbdkp/index.php">avathar.be</a> s’il y a une mise à jour.',
 
-	'PBWOW_CPF_CHECK'					=> 'Vérification des champs de profil personnalisés',
-	//'PBWOW_CPF_CREATE_LOCATION'			=> 'Créer ou activer ces champs depuis le panneau d’administration > Utilisateurs et Groupes > Champs de profil personnalisés',
-	'PBWOW_CPF_LOAD_LOCATION'			=> 'Activer ces champs depuis le panneau d’administration > Général > Configuration Générale > Fonctionnalités du forums',
-	'PBWOW_GAME_EXPLAIN'				=> 'Les champs de profil personnalisés pour ce jeu sont actuellement désactivés.',
-
-	'PBWOW_BNETCHARS_CHECK'				=> 'Fonctionnalités sur les informations des personnages avec l’API de Battle.net',
-	'PBWOW_CHARSDB_GOOD'				=> 'Table des personnages de PBWoW trouvée (%s)',
-	'PBWOW_CHARSDB_BAD'					=> 'Aucune table des personnages de PBWoW n’a été trouvée. S’assurer que la table (%s) existe dans la base de données du forum phpBB.',
-	'PBWOW_CHARSDB_BAD_EXPLAIN'			=> 'La table des personnages Battle.net de PBWoW 3 pour devrait avoir été installée automatiquement lorsque vous avez activé l’extension PBWoW. Veuillez désactiver l’extension, supprimer ses données et essayer de l’activer à nouveau.',
-	'PBWOW_CHARSDB_FLUSH'				=> 'Vider la tables des personnages',
-	'PBWOW_CHARSDB_FLUSH_EXPLAIN'		=> 'Cela effacera toutes les informations sur les personnages Battle.net stockées dans la base de données. Elles seront à nouveau récupérées automatiquement si nécessaire.',
-	'PBWOW_CURL_BAD'					=> 'Votre serveur ne supporte pas &quot;cURL&quot; !',
-	'PBWOW_CURL_BAD_EXPLAIN'			=> 'Vérifier la configuration de votre serveur, ou contacter votre hébergeur. Désactiver l’API de Battle.net API jusqu’à ce que cURL soit activé !',
-
+	'PBWOW_DONATE_URL'                  => 'http://www.avathar.be/bbdkp/app.php/page/donate',
+	'PAYPAL_IMAGE_URL'                  => 'https://www.paypalobjects.com/webstatic/en_US/i/btn/png/silver-pill-paypal-26px.png',
+	'PAYPAL_ALT'                        => 'Donation PayPal',
 	'PBWOW_DONATE'						=> 'Faire un don à PBWoW',
 	'PBWOW_DONATE_SHORT'				=> 'Faire un don à PBWoW',
 	'PBWOW_DONATE_EXPLAIN'				=> 'PBWoW est 100% libre. ce projet est un passe-temps où je consacre mon temps et mon argent, juste pour le plaisir. Si vous appréciez PBWoW, vous pouvez envisager de faire un don. Il sera grandement apprécié. Sans contrepartie.',
-
-
-	// LEGACY CHECKS //
-
-	'PBWOW_LEGACY_CHECK'				=> 'Vérification de l’antériorité de PBWoW',
-
-	'PBWOW_LEGACY_CONSTANTS'			=> 'Vérification des problèmes persistants de PBWoW',
-	'PBWOW_LEGACY_CONSTANTS_EXPLAIN'	=> 'Si détecté, cela signifie qu’il y a des MODs (partiels) de PBWoW v1 ou v2 actifs ! Cela pourrait conduire à des erreurs. C’est pourquoi nous vous recommandons fortement de désactiver tout MOD actif (PBWoW) avant de passer à la dernière version de phpBB. Ou d’installer une version propre de phpBB et d’utiliser la fonction de mise à jour de la base de données de l’installateur de phpBB.',
-	'PBWOW_LEGACY_DATABASE'				=> 'PBWoW Legacy Database(s)',
-	'PBWOW_LEGACY_DATABASE_EXPLAIN'		=> 'La table de configuration de PBWoW v1 ou v2 est toujours active. Ce n’est pas un problème, puisque PBWoW 3 n’interagit pas avec elle. Mais vous pouvez la supprimer la table si vous voulez (elle ne sera plus utilisée).',
-
-	'PBWOW_LEGACY_NONE'					=> 'Aucun problème antérieur concernant des anciennes versions de PBWoW n’a été trouvé. C’est bon signe.',
-
 
 	// CONFIG //
 
@@ -133,24 +108,6 @@ $lang = array_merge($lang, array(
 	'PBWOW_VIDEOBG_ALLPAGES_EXPLAIN'	=> 'Par défaut, PBWoW charge seulement les arrières plans vidéo (si disponibles) sur la page <u>index.php</u>. Vous pouvez étendre l’affichage des arrières plans vidéo sur toutes les pages, mais cela peut affecter la vitesse de navigation de vos visiteurs (pas la bande passante de votre serveur, car ils sont mis en cache localement). [ne s’applique que si la vidéo est activée]',
 	'PBWOW_FIXEDBG'						=> 'Fixer la position de l’arrière plan',
 	'PBWOW_FIXEDBG_EXPLAIN'				=> 'Fixer la position de l’arrière plan permet d’éviter le défilement avec le reste du contenu. Gardez à l’esprit que certains appareils ayant une faible résolution n’auront pas la possibilité de voir l’image d’arrière plan toute entière.',
-
-	'PBWOW_AVATARS'						=> 'Avatars de jeu',
-	'PBWOW_AVATARS_ENABLE'				=> 'Activer le support étendu des avatars de jeu (et des icônes)',
-	'PBWOW_AVATARS_ENABLE_EXPLAIN'		=> 'Si activé, votre forum affichera un avatar de jeu généré (en fonction des champs du profil) si l’utilisateur n’a pas configuré d’avatar personnalisé.',
-	'PBWOW_AVATARS_PATH'				=> 'Chemin des avatars de jeu',
-	'PBWOW_AVATARS_PATH_EXPLAIN'		=> 'Chemin depuis le répertoire racine de votre forum phpBB où les avatars de jeu seront stockés, exemple <samp>images/avatars/gaming</samp>.<br />Les icônes des personnages exigent également ce chemin pour être utilisés.',
-	'PBWOW_SMALLRANKS_ENABLE'			=> 'Utiliser de petites images de rang',
-	'PBWOW_SMALLRANKS_ENABLE_EXPLAIN'	=> 'Activer cette option pour utiliser de petites images de rang qui se superposent à l’avatar (comme c’est le cas sur PBWoW.com). Ne pas activer cette option si vous utilisez de plus grandes images de rang.',
-
-	'PBWOW_BNET_APIKEY'					=> 'Clé de l’API Battle.net',
-	'PBWOW_BNET_APIKEY_EXPLAIN'			=> 'Saisir votre clé du jeu pour l’API Battle.net. Si vous n’en avez pas, il est possible d’en obtenir une en créant un <a href="https://dev.battle.net/member/register">compte Mashery</a>.',
-	'PBWOW_BNETCHARS'					=> 'Informations sur les personnages Battle.net',
-	'PBWOW_BNETCHARS_ENABLE'			=> 'Activer l’API Battle.net pour les informations des personnages',
-	'PBWOW_BNETCHARS_ENABLE_EXPLAIN'	=> 'Activer cette fonctionnalité pour utilise l’API Battle.net afin de récupérer les informations des personnages (lorsqu’elles sont disponibles), pour les utiliser dans les profils d’utilisateur. Le paramètre des <u>avatars de jeu</u> doit être activé pour afficher les avatars Battle.net !',
-	'PBWOW_BNETCHARS_CACHETIME'			=> 'Durée de vie du cache',
-	'PBWOW_BNETCHARS_CACHETIME_EXPLAIN'	=> 'Paramétrer la durée de vie du cache (en secondes) des informations des personnages après qu’elles aient été récupérées à partir de l’API Battle.net. Vous pouvez modifier ce paramètre pour mettre à jour les informations des personnages plus ou moins fréquemment. 86400 = 24h',
-	'PBWOW_BNETCHARS_TIMEOUT'			=> 'Délai de la requête de l’API',
-	'PBWOW_BNETCHARS_TIMEOUT_EXPLAIN'	=> 'Définir le temps écoulé limite (en secondes) pour les requêtes vers l’API Battle.net. Il s’agit du temps maximum que le script attendra avant que Battle.net réponde. Augmenter la valeur si les données (correctes) ne sont pas reçues à temps, mais le temps de chargement de la page peut augmenter !',
 
 	'PBWOW_ADS_INDEX'					=> 'Encart publicitaire sur l’index',
 	'PBWOW_ADS_INDEX_ENABLE'			=> 'Activer l’encart publicitaire sur l’index',